+++ Editor/ScriptRunner/ScriptRunner.cs
@@ -14,15 +14,18 @@ namespace SuperUnityBuild.BuildActions
 
         public override void Execute()
         {
-            RunScript(scriptPath, scriptArguments);
+            string resolvedScriptPath = BuildAction.ResolveExecuteTokens(scriptPath);
+            string resolvedScriptArguments = BuildAction.ResolveExecuteTokens(scriptArguments);
+
+            RunScript(resolvedScriptPath, resolvedScriptArguments);
         }
 
         public override void PerBuildExecute(BuildReleaseType releaseType, BuildPlatform platform, BuildArchitecture architecture, BuildScriptingBackend scriptingBackend, BuildDistribution distribution, DateTime buildTime, ref BuildOptions options, string configKey, string buildPath)
         {
-            string resolvedScriptPath = BuildProject.ResolvePath(scriptPath, releaseType, platform, architecture, scriptingBackend, distribution, buildTime);
-            string resolvedScriptArgs = BuildProject.ResolvePath(scriptArguments, releaseType, platform, architecture, scriptingBackend, distribution, buildTime);
+            string resolvedScriptPath = BuildAction.ResolvePerBuildExecuteTokens(scriptPath, releaseType, platform, architecture, scriptingBackend, distribution, buildTime, buildPath);
+            string resolvedScriptArguments = BuildAction.ResolvePerBuildExecuteTokens(scriptArguments, releaseType, platform, architecture, scriptingBackend, distribution, buildTime, buildPath);
 
-            RunScript(resolvedScriptPath, resolvedScriptArgs);
+            RunScript(resolvedScriptPath, resolvedScriptArguments);
         }
 
         private void RunScript(string scriptPath, string arguments)